fix: remove metadata from client components
